HAAR-827: Use get Role Eetails endpoint in external users api - instead of auth
