+++ src/main/kotlin/uk/gov/justice/digital/hmpps/manageusersapi/service/ExternalUsersApiService.kt
@@ -3,13 +3,21 @@ package uk.gov.justice.digital.hmpps.manageusersapi.service
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Qualifier
+import org.springframework.beans.factory.annotation.Value
+import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import org.springframework.web.reactive.function.client.WebClient
+import org.springframework.web.reactive.function.client.WebClientResponseException
 import uk.gov.justice.digital.hmpps.manageusersapi.resource.Role
 import uk.gov.justice.digital.hmpps.manageusersapi.resource.RolesPaged
+import java.time.Duration
 
 @Service
-class ExternalUsersApiService(@Qualifier("externalUsersWebClient") val externalUsersWebClient: WebClient) {
+class ExternalUsersApiService(
+  @Qualifier("externalUsersWebClient") val externalUsersWebClient: WebClient,
+  @Value("\${api.timeout:10s}")
+  val timeout: Duration
+) {
   companion object {
     val log: Logger = LoggerFactory.getLogger(this::class.java)
   }
@@ -51,6 +59,19 @@ class ExternalUsersApiService(@Qualifier("externalUsersWebClient") val externalU
       .retrieve()
       .bodyToMono(RolesPaged::class.java)
       .block()!!
+  }
+
+  @Throws(RoleNotFoundException::class)
+  fun getRoleDetail(roleCode: String): Role {
+    try {
+      return externalUsersWebClient.get()
+        .uri("/roles/$roleCode")
+        .retrieve()
+        .bodyToMono(Role::class.java)
+        .block(timeout) ?: throw RoleNotFoundException("get", roleCode, "notfound")
+    } catch (e: WebClientResponseException) {
+      throw if (e.statusCode.equals(HttpStatus.NOT_FOUND)) RoleNotFoundException("get", roleCode, "notfound") else e
+    }
   }
 }
