+++ src/test/kotlin/uk/gov/justice/digital/hmpps/manageusersapi/integration/wiremock/HmppsAuthMockServer.kt
@@ -43,69 +43,6 @@ class HmppsAuthMockServer : WireMockServer(WIREMOCK_PORT) {
     )
   }
 
-  fun stubGetRolesDetails(roleCode: String) {
-    stubFor(
-      get(urlEqualTo("/auth/api/roles/$roleCode"))
-        .willReturn(
-          aResponse()
-            .withHeaders(HttpHeaders(HttpHeader("Content-Type", "application/json")))
-            .withBody(
-              """{
-                    "roleCode": "$roleCode",
-                    "roleName": "Group Manager",
-                    "roleDescription": "Allow Group Manager to administer the account within their groups",
-                    "adminType": [
-                        {
-                            "adminTypeCode": "EXT_ADM",
-                            "adminTypeName": "External Administrator"
-                        }
-                    ]
-                  }
-              """.trimIndent()
-            )
-        )
-    )
-  }
-
-  fun stubGetDPSRoleDetails(roleCode: String) {
-    stubFor(
-      get(urlEqualTo("/auth/api/roles/$roleCode"))
-        .willReturn(
-          aResponse()
-            .withHeaders(HttpHeaders(HttpHeader("Content-Type", "application/json")))
-            .withBody(
-              """{
-                    "roleCode": "$roleCode",
-                    "roleName": "Group Manager",
-                    "roleDescription": "Allow Group Manager to administer the account within their groups",
-                    "adminType": [
-                        {
-                            "adminTypeCode": "EXT_ADM",
-                            "adminTypeName": "External Administrator"
-                        },
-                        {
-                            "adminTypeCode": "DPS_ADM",
-                            "adminTypeName": "DPS Central Administrator"
-                        }
-                    ]
-                  }
-              """.trimIndent()
-            )
-        )
-    )
-  }
-
-  fun stubGetRoleDetailsFail(status: HttpStatus, roleCode: String) {
-    stubFor(
-      get(urlEqualTo("/auth/api/roles/$roleCode"))
-        .willReturn(
-          aResponse()
-            .withHeader("Content-Type", "application/json")
-            .withStatus(status.value())
-        )
-    )
-  }
-
   fun stubCreateRole() {
     stubFor(
       post(urlEqualTo("/auth/api/roles"))
